Refactor opening of nethuns socket
- Remove the factory struct for Nethuns sockets and convert the code to a more C-like approach (a simple function called `nethuns_socket_open`)
- Remove the `open` function from the `BindableNethunsSocket` trait, since it wasn't useful beacuse of the "factory" pattern
- Changes to the exported API in `lib.rs` for better cleanness
