+++ src/misc.rs
@@ -2,8 +2,8 @@ mod circular_buffer;
 
 pub(crate) use circular_buffer::*;
 
+use crate::sockets::NethunsSocket;
 use crate::types::NethunsQueue;
-use crate::NethunsSocket;
 
 
 /// Get full device name, taking into account
