+++ src/sockets.rs
@@ -17,19 +17,19 @@ use self::errors::{
 
 /*
     Import the structs defined for the required I/O framework.
-
+    
     Every framework-specific implementation must provide:
-    - A struct which implements the `NethunsSocket` trait, which will
-        be built by the `NethunsSocketFactory` factory.
+    - A struct which implements the `BindableNethunsSocket` trait.
+    - A struct which implements the `NethunsSocket` trait.
     - A struct named `Pkthdr` which must implement the `PkthdrTrait` trait.
     TODO: move it to mod documentation
 */
 cfg_if::cfg_if! {
     if #[cfg(feature="netmap")] {
         mod netmap;
-
+        
         use netmap::bindable_socket::BindableNethunsSocketNetmap;
-
+        
         pub use netmap::pkthdr::Pkthdr;
     }
     else {
@@ -38,27 +38,36 @@ cfg_if::cfg_if! {
 }
 
 
-/// Trait which defines the interface for a Nethuns socket 
+/// Open a new Nethuns socket, by calling the `open` function
+/// of the struct belonging to the I/O framework selected at compile time.
+///
+/// # Arguments
+/// * `opt`: The options for the socket.
+///
+/// # Returns
+/// * `Ok(Box<dyn BindableNethunsSocket>)` - A new nethuns socket, in no error occurs.
+/// * `Err(NethunsOpenError::InvalidOptions)` - If at least one of the options holds a invalid value.
+/// * `Err(NethunsOpenError::Error)` - If an unexpected error occurs.
+pub fn nethuns_socket_open(
+    opt: NethunsSocketOptions,
+) -> Result<Box<dyn BindableNethunsSocket>, NethunsOpenError> {
+    cfg_if::cfg_if! {
+        if #[cfg(feature="netmap")] {
+            return BindableNethunsSocketNetmap::open(opt);
+        }
+        else {
+            std::compile_error!("The support for the specified I/O framework is not available yet. Check the documentation for more information.");
+        }
+    }
+}
+
+
+/// Trait which defines the interface for a Nethuns socket
 /// not binded to a specific device and queue.
-/// 
+///
 /// In order to properly use the socket, you need to bind it first
 /// to a specific device and queue by calling [`BindableNethunsSocket::bind()`].
 pub trait BindableNethunsSocket: Debug {
-    /// Tries to open a new Nethuns socket.
-    ///
-    /// # Arguments
-    /// * `opt`: The options for the socket.
-    ///
-    /// # Returns
-    /// * `Ok(Box<dyn NethunsSocket>)` - A new nethuns socket, if no error occurred.
-    /// * `Err(NethunsOpenError::InvalidOptions)` - If at least one of the options holds a invalid value.
-    /// * `Err(NethunsOpenError::Error)` - If an unexpected error occurs.
-    fn open(
-        opt: NethunsSocketOptions,
-    ) -> Result<Box<dyn BindableNethunsSocket>, NethunsOpenError>
-    where
-        Self: Sized;
-
     /// Bind an opened socket to a specific queue / any queue of interface/device `dev`.
     ///
     /// # Returns
@@ -74,18 +83,18 @@ pub trait BindableNethunsSocket: Debug {
         Box<dyn NethunsSocket>,
         (NethunsBindError, Box<dyn BindableNethunsSocket>),
     >;
-
+    
     /// Get an immutable reference to the base descriptor of the socket.
     fn base(&self) -> &NethunsSocketBase;
     /// Get an mutable reference to the base descriptor of the socket.
     fn base_mut(&mut self) -> &mut NethunsSocketBase;
-
+    
     /// Check if the socket is in RX mode
     #[inline(always)]
     fn rx(&self) -> bool {
         self.base().rx_ring.is_some()
     }
-
+    
     /// Check if the socket is in TX mode
     #[inline(always)]
     fn tx(&self) -> bool {
@@ -94,7 +103,7 @@ pub trait BindableNethunsSocket: Debug {
 }
 
 
-/// Trait which defines the interface for a Nethuns socket after binding. 
+/// Trait which defines the interface for a Nethuns socket after binding.
 /// This socket is usable for RX and/or TX, depending from its configuration.
 pub trait NethunsSocket: Debug {
     /// Get the next unprocessed received packet.
@@ -108,8 +117,8 @@ pub trait NethunsSocket: Debug {
     /// * `Err(NethunsRecvError::FrameworkError)` - If an error from the unsafe interaction with underlying I/O framework occurs.
     /// * `Err(NethunsRecvError::Error)` - If an unexpected error occurs.
     fn recv(&mut self) -> Result<RecvPacket, NethunsRecvError>;
-
-
+    
+    
     /// Queue up a packet for transmission.
     ///
     /// # Returns
@@ -117,8 +126,8 @@ pub trait NethunsSocket: Debug {
     /// * `Err(NethunsSendError::NotTx)` -  If the socket is not configured in TX mode. Check the configuration parameters passed to `open(...)`.
     /// * `Err(NethunsSendError::InUse)` - If the slot at the tail of the TX ring is not released yet and it's currently in use by the application.
     fn send(&mut self, packet: &[u8]) -> Result<(), NethunsSendError>;
-
-
+    
+    
     /// Send all queued up packets.
     ///
     /// # Returns
@@ -127,8 +136,8 @@ pub trait NethunsSocket: Debug {
     /// * `Err(NethunsFlushError::FrameworkError)` - If an error from the unsafe interaction with underlying I/O framework occurs.
     /// * `Err(NethunsFlushError::Error)` - If an unexpected error occurs.
     fn flush(&mut self) -> Result<(), NethunsFlushError>;
-
-
+    
+    
     /// Mark the packet contained in the a specific slot
     /// of the TX ring as *ready for transmission*.
     ///
@@ -140,31 +149,31 @@ pub trait NethunsSocket: Debug {
     /// * `Ok(())` - On success.
     /// * `Err(NethunsSendError::InUse)` - If the slot is not released yet and it's currently in use by the application.
     fn send_slot(&self, id: usize, len: usize) -> Result<(), NethunsSendError>;
-
-
+    
+    
     /// Get an immutable reference to the base socket descriptor.
     fn base(&self) -> &NethunsSocketBase;
     /// Get a mutable reference to the base socket descriptor.
     fn base_mut(&mut self) -> &mut NethunsSocketBase;
-
-
+    
+    
     /// Check if the socket is in TX mode
     #[inline(always)]
     fn tx(&self) -> bool {
         self.base().tx_ring.is_some()
     }
-
+    
     /// Check if the socket is in RX mode
     #[inline(always)]
     fn rx(&self) -> bool {
         self.base().rx_ring.is_some()
     }
-
-
+    
+    
     /// Get the file descriptor of the socket.
     fn fd(&self) -> libc::c_int;
-
-
+    
+    
     /// Get a mutable reference to the buffer inside
     /// a specific ring slot which will contain the packet
     /// to be sent.
@@ -178,51 +187,22 @@ pub trait NethunsSocket: Debug {
     /// * `Some(&mut [u8])` - buffer reference.
     /// * `None` - if the socket is not in TX mode.
     fn get_packet_buffer_ref(&self, pktid: usize) -> Option<&mut [u8]>;
-
-
+    
+    
     /// Join a fanout group.
     ///
     /// # Arguments
     /// * `group` - The group id.
     /// * `fanout` - A string encoding the details of the fanout mode.
     fn fanout(&mut self, group: libc::c_int, fanout: &CStr) -> bool;
-
-
+    
+    
     /// Dump the rings.
     fn dump_rings(&mut self);
-
+    
     /// Get some statistics about the socket
     /// or `None` on error.
     fn stats(&self) -> Option<NethunsStat>;
-}
-
-
-/// Factory to build objects which implements the trait NethunsSocket
-pub struct NethunsSocketFactory();
-
-impl NethunsSocketFactory {
-    /// Tries to open a new Nethuns socket, by calling the `open` function
-    /// of the struct belonging to the I/O framework selected at compile time.
-    ///
-    /// # Arguments
-    /// * `opt`: The options for the socket.
-    ///
-    /// # Returns
-    /// * `Ok(Box<dyn NethunsSocket>)` - A new nethuns socket, in no error occurs.
-    /// * `Err(NethunsOpenError::InvalidOptions)` - If at least one of the options holds a invalid value.
-    /// * `Err(NethunsOpenError::Error)` - If an unexpected error occurs.
-    pub fn nethuns_socket_open(
-        opt: NethunsSocketOptions,
-    ) -> Result<Box<dyn BindableNethunsSocket>, NethunsOpenError> {
-        cfg_if::cfg_if! {
-            if #[cfg(feature="netmap")] {
-                return BindableNethunsSocketNetmap::open(opt);
-            }
-            else {
-                std::compile_error!("The support for the specified I/O framework is not available yet. Check the documentation for more information.");
-            }
-        }
-    }
 }
 
 
@@ -236,14 +216,14 @@ pub trait PkthdrTrait: Debug {
     fn tstamp_set_sec(&mut self, sec: u32);
     fn tstamp_set_usec(&mut self, usec: u32);
     fn tstamp_set_nsec(&mut self, nsec: u32);
-
+    
     fn snaplen(&self) -> u32;
     fn len(&self) -> u32;
     fn set_snaplen(&mut self, len: u32);
     fn set_len(&mut self, len: u32);
-
+    
     fn rxhash(&self) -> u32;
-
+    
     fn offvlan_tpid(&self) -> u16;
     fn offvlan_tci(&self) -> u16;
 }
@@ -252,7 +232,7 @@ pub trait PkthdrTrait: Debug {
 #[cfg(test)]
 mod test {
     use is_trait::is_trait;
-
+    
     #[test]
     /// Make sure that the NethunsSocket trait is implemented correctly.
     fn assert_nethuns_socket_trait() {
@@ -260,7 +240,7 @@ mod test {
             if #[cfg(feature="netmap")] {
                 assert!(
                     is_trait!(
-                        super::netmap::nethuns_socket::NethunsSocketNetmap, 
+                        super::netmap::nethuns_socket::NethunsSocketNetmap,
                         super::NethunsSocket
                     )
                 );
